Partition split-up binding code between source and target files
Final step in splitting code between source and target files.

  - Remove moved code from source.
  - Remove unmoved code from target.
  - Add some infrastructure to target in preparation to build.
  - Add target files to build infrastructure.
