+++ bindings/pydrake/geometry_py_scene_graph.cc
@@ -1,23 +1,14 @@
-#include <sstream>
+/* @file This includes the SceneGraph class and the major components of its
+ API: SceneGraphInspector and QueryObject for examining its state and performing
+ queries, and the query result types as well. They can be found in the
+ pydrake.geometry module. */
 
-#include "pybind11/eigen.h"
-#include "pybind11/eval.h"
-#include "pybind11/operators.h"
-#include "pybind11/pybind11.h"
-#include "pybind11/stl.h"
-
-#include "drake/bindings/pydrake/common/cpp_template_pybind.h"
 #include "drake/bindings/pydrake/common/default_scalars_pybind.h"
 #include "drake/bindings/pydrake/common/deprecation_pybind.h"
 #include "drake/bindings/pydrake/common/type_pack.h"
-#include "drake/bindings/pydrake/common/type_safe_index_pybind.h"
 #include "drake/bindings/pydrake/common/value_pybind.h"
 #include "drake/bindings/pydrake/documentation_pybind.h"
-#include "drake/bindings/pydrake/geometry_py.h"
-#include "drake/bindings/pydrake/pydrake_pybind.h"
 #include "drake/geometry/geometry_frame.h"
-#include "drake/geometry/proximity_properties.h"
-#include "drake/geometry/query_results/penetration_as_point_pair.h"
 #include "drake/geometry/scene_graph.h"
 
 // TODO(SeanCurtis-TRI) When pybind issue 3019 gets resolved, we won't need to
@@ -43,16 +34,8 @@ namespace drake {
 namespace pydrake {
 namespace {
 
-using Eigen::Vector3d;
-using geometry::GeometryId;
-using geometry::PerceptionProperties;
-using geometry::Shape;
-using math::RigidTransformd;
 using systems::Context;
 using systems::LeafSystem;
-using systems::sensors::ImageDepth32F;
-using systems::sensors::ImageLabel16I;
-using systems::sensors::ImageRgba8U;
 
 template <typename T>
 void DoScalarDependentDefinitions(py::module m, T) {
@@ -492,6 +475,38 @@ void DoScalarDependentDefinitions(py::module m, T) {
     AddValueInstantiation<QueryObject<T>>(m);
   }
 
+  // ContactSurface
+  {
+    using Class = ContactSurface<T>;
+    auto cls = DefineTemplateClassWithDefault<Class>(
+        m, "ContactSurface", param, doc.ContactSurface.doc);
+    cls  // BR
+        .def("id_M", &Class::id_M, doc.ContactSurface.id_M.doc)
+        .def("id_N", &Class::id_N, doc.ContactSurface.id_N.doc)
+        .def("mesh_W", &Class::mesh_W, doc.ContactSurface.mesh_W.doc);
+  }
+
+  // PenetrationAsPointPair
+  {
+    using Class = PenetrationAsPointPair<T>;
+    auto cls = DefineTemplateClassWithDefault<Class>(
+        m, "PenetrationAsPointPair", param, doc.PenetrationAsPointPair.doc);
+    cls  // BR
+        .def(ParamInit<Class>(), doc.PenetrationAsPointPair.ctor.doc)
+        .def_readwrite("id_A", &PenetrationAsPointPair<T>::id_A,
+            doc.PenetrationAsPointPair.id_A.doc)
+        .def_readwrite("id_B", &PenetrationAsPointPair<T>::id_B,
+            doc.PenetrationAsPointPair.id_B.doc)
+        .def_readwrite("p_WCa", &PenetrationAsPointPair<T>::p_WCa,
+            py::return_value_policy::copy, doc.PenetrationAsPointPair.p_WCa.doc)
+        .def_readwrite("p_WCb", &PenetrationAsPointPair<T>::p_WCb,
+            py::return_value_policy::copy, doc.PenetrationAsPointPair.p_WCb.doc)
+        .def_readwrite("nhat_BA_W", &PenetrationAsPointPair<T>::nhat_BA_W,
+            doc.PenetrationAsPointPair.nhat_BA_W.doc)
+        .def_readwrite("depth", &PenetrationAsPointPair<T>::depth,
+            doc.PenetrationAsPointPair.depth.doc);
+  }
+
   // SignedDistancePair
   {
     using Class = SignedDistancePair<T>;
@@ -534,72 +549,13 @@ void DoScalarDependentDefinitions(py::module m, T) {
             return_value_policy_for_scalar_type<T>(),
             doc.SignedDistanceToPoint.grad_W.doc);
   }
-
-  // PenetrationAsPointPair
-  {
-    using Class = PenetrationAsPointPair<T>;
-    auto cls = DefineTemplateClassWithDefault<Class>(
-        m, "PenetrationAsPointPair", param, doc.PenetrationAsPointPair.doc);
-    cls  // BR
-        .def(ParamInit<Class>(), doc.PenetrationAsPointPair.ctor.doc)
-        .def_readwrite("id_A", &PenetrationAsPointPair<T>::id_A,
-            doc.PenetrationAsPointPair.id_A.doc)
-        .def_readwrite("id_B", &PenetrationAsPointPair<T>::id_B,
-            doc.PenetrationAsPointPair.id_B.doc)
-        .def_readwrite("p_WCa", &PenetrationAsPointPair<T>::p_WCa,
-            py::return_value_policy::copy, doc.PenetrationAsPointPair.p_WCa.doc)
-        .def_readwrite("p_WCb", &PenetrationAsPointPair<T>::p_WCb,
-            py::return_value_policy::copy, doc.PenetrationAsPointPair.p_WCb.doc)
-        .def_readwrite("nhat_BA_W", &PenetrationAsPointPair<T>::nhat_BA_W,
-            doc.PenetrationAsPointPair.nhat_BA_W.doc)
-        .def_readwrite("depth", &PenetrationAsPointPair<T>::depth,
-            doc.PenetrationAsPointPair.depth.doc);
-  }
-
-  // ContactSurface
-  {
-    using Class = ContactSurface<T>;
-    auto cls = DefineTemplateClassWithDefault<Class>(
-        m, "ContactSurface", param, doc.ContactSurface.doc);
-    cls  // BR
-        .def("id_M", &Class::id_M, doc.ContactSurface.id_M.doc)
-        .def("id_N", &Class::id_N, doc.ContactSurface.id_N.doc)
-        .def("mesh_W", &Class::mesh_W, doc.ContactSurface.mesh_W.doc);
-  }
-}  // NOLINT(readability/fn_size)
-
-void def_geometry(py::module m) {
-  type_visit([m](auto dummy) { DoScalarDependentDefinitions(m, dummy); },
-      NonSymbolicScalarPack{});
-}
-
-void def_geometry_all(py::module m) {
-  py::dict vars = m.attr("__dict__");
-  py::exec(
-      "from pydrake.geometry import *\n"
-      "from pydrake.geometry.render import *\n"
-      "from pydrake.geometry.optimization import *\n",
-      py::globals(), vars);
-}
-
-PYBIND11_MODULE(geometry, m) {
-  PYDRAKE_PREVENT_PYTHON3_MODULE_REIMPORT(m);
-  py::module::import("pydrake.common");
-  py::module::import("pydrake.math");
-  py::module::import("pydrake.systems.framework");
-  py::module::import("pydrake.systems.lcm");
-
-  /* The order of execution matters -- a module may rely on the definition
-   of bindings executed prior to it. */
-  DefineGeometryCommon(m);
-  DefineGeometryHydro(m);
-  def_geometry(m);
-  DefineGeometryRender(m.def_submodule("render"));
-  DefineGeometryOptimization(m.def_submodule("optimization"));
-  DefineGeometryVisualizers(m);
-  def_geometry_all(m.def_submodule("all"));
 }
 
 }  // namespace
+
+void DefineGeometrySceneGraph(py::module m) {
+  type_visit([m](auto dummy) { DoScalarDependentDefinitions(m, dummy); },
+      NonSymbolicScalarPack{});
+}
 }  // namespace pydrake
 }  // namespace drake
